@@ -2,11 +2,11 @@
 
   <!-- MSBuild properties -->
   <PropertyGroup>
-    <RootNamespace>c2ffi.Tool</RootNamespace>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net8.0</TargetFramework>
+    <TargetFramework>net9.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
+    <RootNamespace>c2ffi</RootNamespace>
   </PropertyGroup>
 
   <!-- NuGet -->
@@ -27,31 +27,31 @@
 
   <!-- Read .csproj information at runtime -->
   <ItemGroup>
-    <AssemblyAttribute Include="c2ffi.Tool.ProjectInfoAttribute">
+    <AssemblyAttribute Include="c2ffi.ProjectInfoAttribute">
       <_Parameter1>$(PackageDescription)</_Parameter1>
     </AssemblyAttribute>
   </ItemGroup>
 
   <!-- NuGet package references -->
   <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="8.0.0" />
-    <PackageReference Include="Microsoft.Extensions.Hosting" Version="8.0.0" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="9.0.0" />
+    <PackageReference Include="Microsoft.Extensions.Hosting" Version="9.0.0" />
     <PackageReference Include="System.CommandLine" Version="2.0.0-beta4.22272.1" />
     <PackageReference Include="System.IO.Abstractions" Version="20.0.15" />
     <PackageReference Include="bottlenoselabs.Common" Version="2024.4.11" />
+  </ItemGroup>
+
+  <!-- Project references -->
+  <ItemGroup>
+    <ProjectReference Include="../c2ffi.Data/c2ffi.Data.csproj" />
   </ItemGroup>
 
   <!-- Configuration settings -->
   <ItemGroup>
     <EmbeddedResource Include="Properties/appsettings.json" />
   </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\c2ffi.Data\c2ffi.Data.csproj" />
-  </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Commands\Extract\Infrastructure\" />
-    <Folder Include="Generated\" />
-  </ItemGroup>
+
+  <!-- Generated source code files -->
   
   <!-- libclang -->
   <ItemGroup>
