@@ -2,7 +2,7 @@
 
   <!-- MSBuild project settings -->
   <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
+    <TargetFramework>net9.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
     <IsPackable>false</IsPackable>
@@ -12,15 +12,15 @@
 
   <!-- NuGet package references -->
   <ItemGroup>
-    <PackageReference Include="FluentAssertions" Version="6.12.0" />
-    <PackageReference Include="JetBrains.Annotations" Version="2023.3.0" />
-    <PackageReference Include="xunit" Version="2.4.2"/>
-    <PackageReference Include="System.IO.Abstractions" Version="20.0.15" />
+    <PackageReference Include="FluentAssertions" Version="7.0.0" />
+    <PackageReference Include="JetBrains.Annotations" Version="2024.3.0" />
+    <PackageReference Include="xunit" Version="2.9.2"/>
+    <PackageReference Include="System.IO.Abstractions" Version="21.1.7" />
   </ItemGroup>
 
   <!-- Project references -->
   <ItemGroup>
-    <ProjectReference Include="..\..\production\c2ffi.Tool\c2ffi.Tool.csproj" />
+    <ProjectReference Include="../../production/c2ffi.Tool/c2ffi.Tool.csproj" />
   </ItemGroup>
 
 </Project>
