@@ -1,7 +1,6 @@
 // Copyright (c) Bottlenose Labs Inc. (https://github.com/bottlenoselabs). All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the Git repository root directory for full license information.
 
-using System.Diagnostics.CodeAnalysis;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 
@@ -46,6 +45,6 @@ public class TargetPlatformJsonConverter : JsonConverter<TargetPlatform>
         TargetPlatform value,
         JsonSerializerOptions options)
     {
-        writer.WriteStringValue(value.ClangTargetTriple);
+        writer.WriteStringValue(value._clangTargetTriple);
     }
 }
