@@ -7,36 +7,31 @@ using FluentAssertions.Primitives;
 
 namespace c2ffi.Tests.Library.Assertions;
 
-public class CTestTypeAssertions : ReferenceTypeAssertions<CTestType?, CTestTypeAssertions>
+public class CTestTypeAssertions(CTestType? instance) : ReferenceTypeAssertions<CTestType?, CTestTypeAssertions>(instance)
 {
-    public CTestTypeAssertions(CTestType? instance)
-        : base(instance)
-    {
-    }
-
     protected override string Identifier => "type";
 
     [CustomAssertion]
     public void BeChar(string because = "", params object[] becauseArgs)
     {
-        Subject.Should().NotBeNull(because, becauseArgs);
-        Subject!.Name.Should().Be("char", because, becauseArgs);
-        Subject.NodeKind.Should().Be("primitive", because, becauseArgs);
-        Subject.AlignOf.Should().Be(1, because, becauseArgs);
-        Subject.SizeOf.Should().Be(1, because, becauseArgs);
-        Subject.IsAnonymous.Should().Be(false, because, becauseArgs);
-        Subject.InnerType.Should().BeNull(because, becauseArgs);
+        _ = Subject.Should().NotBeNull(because, becauseArgs);
+        _ = Subject!.Name.Should().Be("char", because, becauseArgs);
+        _ = Subject.NodeKind.Should().Be("primitive", because, becauseArgs);
+        _ = Subject.AlignOf.Should().Be(1, because, becauseArgs);
+        _ = Subject.SizeOf.Should().Be(1, because, becauseArgs);
+        _ = Subject.IsAnonymous.Should().Be(false, because, becauseArgs);
+        _ = Subject.InnerType.Should().BeNull(because, becauseArgs);
     }
 
     [CustomAssertion]
     public void BeInt(string because = "", params object[] becauseArgs)
     {
-        Subject.Should().NotBeNull(because, becauseArgs);
-        Subject!.Name.Should().Be("int", because, becauseArgs);
-        Subject.NodeKind.Should().Be("primitive", because, becauseArgs);
-        Subject.AlignOf.Should().Be(4, because, becauseArgs);
-        Subject.SizeOf.Should().Be(4, because, becauseArgs);
-        Subject.IsAnonymous.Should().Be(false, because, becauseArgs);
-        Subject.InnerType.Should().BeNull(because, becauseArgs);
+        _ = Subject.Should().NotBeNull(because, becauseArgs);
+        _ = Subject!.Name.Should().Be("int", because, becauseArgs);
+        _ = Subject.NodeKind.Should().Be("primitive", because, becauseArgs);
+        _ = Subject.AlignOf.Should().Be(4, because, becauseArgs);
+        _ = Subject.SizeOf.Should().Be(4, because, becauseArgs);
+        _ = Subject.IsAnonymous.Should().Be(false, because, becauseArgs);
+        _ = Subject.InnerType.Should().BeNull(because, becauseArgs);
     }
 }
