@@ -2,7 +2,6 @@
 // Licensed under the MIT license. See LICENSE file in the Git repository root directory for full license information.
 
 using c2ffi.Tests.Library.Helpers;
-using c2ffi.Tool;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
@@ -27,6 +26,6 @@ public static class TestHost
 
     private static void ConfigureTestServices(IServiceCollection services)
     {
-        services.AddSingleton<FileSystemHelper>();
+        _ = services.AddSingleton<FileSystemHelper>();
     }
 }
