@@ -1,9 +1,6 @@
 // Copyright (c) Bottlenose Labs Inc. (https://github.com/bottlenoselabs). All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the Git repository root directory for full license information.
 
-using c2ffi.Tests.Library.Models;
-using FluentAssertions;
-
 #pragma warning disable CA1707
 
 namespace c2ffi.Tests.EndToEnd.Extract.Enums.enum_uint8;
@@ -29,12 +26,12 @@ public class Test : ExtractFfiTest
     {
         const string name = $"enum {EnumName}";
         var @enum = ffi.GetEnum(name);
-        @enum.SizeOf.Should().Be(4);
+        _ = @enum.SizeOf.Should().Be(4);
 
-        @enum.Values[0].Name.Should().Be("ENUM_UINT8_MIN");
-        @enum.Values[0].Value.Should().Be(0);
+        _ = @enum.Values[0].Name.Should().Be("ENUM_UINT8_MIN");
+        _ = @enum.Values[0].Value.Should().Be(0);
 
-        @enum.Values[1].Name.Should().Be("ENUM_UINT8_MAX");
-        @enum.Values[1].Value.Should().Be(255);
+        _ = @enum.Values[1].Name.Should().Be("ENUM_UINT8_MAX");
+        _ = @enum.Values[1].Value.Should().Be(255);
     }
 }
