@@ -5,12 +5,9 @@ using System.Collections.Immutable;
 using System.Diagnostics.CodeAnalysis;
 using System.IO.Abstractions;
 using c2ffi.Data;
-using c2ffi.Data.Nodes;
 using c2ffi.Data.Serialization;
 using c2ffi.Tests.Library;
 using c2ffi.Tests.Library.Helpers;
-using c2ffi.Tests.Library.Models;
-using c2ffi.Tool.Commands.Extract;
 using JetBrains.Annotations;
 using Microsoft.Extensions.DependencyInjection;
 
@@ -25,7 +22,7 @@ public abstract class ExtractFfiTest
     private readonly IDirectory _directory;
     private readonly IFile _file;
     private readonly FileSystemHelper _fileSystemHelper;
-    private readonly ExtractFfiTool _tool;
+    private readonly c2ffi.Extract.Tool _tool;
 
     protected ExtractFfiTest()
     {
@@ -35,7 +32,7 @@ public abstract class ExtractFfiTest
         _directory = _fileSystem.Directory;
         _file = _fileSystem.File;
         _fileSystemHelper = services.GetService<FileSystemHelper>()!;
-        _tool = services.GetService<ExtractFfiTool>()!;
+        _tool = services.GetService<c2ffi.Extract.Tool>()!;
     }
 
     public ImmutableArray<CTestFfiTargetPlatform> GetTargetPlatformFfis(string configurationFilePath)
