@@ -1,9 +1,6 @@
 // Copyright (c) Bottlenose Labs Inc. (https://github.com/bottlenoselabs). All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the Git repository root directory for full license information.
 
-using c2ffi.Tests.Library.Models;
-using FluentAssertions;
-
 #pragma warning disable CA1308
 #pragma warning disable CA1707
 
@@ -29,6 +26,6 @@ public class Test : ExtractFfiTest
     private void FfiMacroObjectDoesNotExist(CTestFfiTargetPlatform ffi)
     {
         var macroObject = ffi.TryGetMacroObject(MacroObjectName);
-        macroObject.Should().Be(null);
+        _ = macroObject.Should().Be(null);
     }
 }
