@@ -26,84 +26,84 @@ public class Test : ExtractFfiTest
     {
         const string name = $"struct {StructName}";
         var @struct = ffi.GetRecord(name);
-        @struct.Name.Should().Be(name);
-        @struct.IsStruct.Should().BeTrue();
-        @struct.IsUnion.Should().BeFalse();
-        @struct.IsAnonymous.Should().BeFalse();
+        _ = @struct.Name.Should().Be(name);
+        _ = @struct.IsStruct.Should().BeTrue();
+        _ = @struct.IsUnion.Should().BeFalse();
+        _ = @struct.IsAnonymous.Should().BeFalse();
 
-        @struct.Fields.Length.Should().Be(1);
+        _ = @struct.Fields.Length.Should().Be(1);
 
         var field = @struct.Fields[0];
-        field.Name.Should().BeEmpty();
-        field.OffsetOf.Should().Be(0);
+        _ = field.Name.Should().BeEmpty();
+        _ = field.OffsetOf.Should().Be(0);
 
         var fieldType = field.Type;
-        fieldType.Name.Should().Be(name + "_ANONYMOUS_0");
-        fieldType.SizeOf.Should().Be(16);
-        fieldType.AlignOf.Should().Be(4);
-        fieldType.IsAnonymous.Should().BeTrue();
-        fieldType.InnerType.Should().BeNull();
+        _ = fieldType.Name.Should().Be(name + "_ANONYMOUS_0");
+        _ = fieldType.SizeOf.Should().Be(16);
+        _ = fieldType.AlignOf.Should().Be(4);
+        _ = fieldType.IsAnonymous.Should().BeTrue();
+        _ = fieldType.InnerType.Should().BeNull();
 
         var anonymousStruct = ffi.GetRecord(fieldType.Name);
-        anonymousStruct.IsStruct.Should().BeTrue();
-        anonymousStruct.IsUnion.Should().BeFalse();
-        anonymousStruct.SizeOf.Should().Be(16);
-        anonymousStruct.AlignOf.Should().Be(4);
-        anonymousStruct.IsAnonymous.Should().BeTrue();
-        anonymousStruct.Fields.Length.Should().Be(2);
+        _ = anonymousStruct.IsStruct.Should().BeTrue();
+        _ = anonymousStruct.IsUnion.Should().BeFalse();
+        _ = anonymousStruct.SizeOf.Should().Be(16);
+        _ = anonymousStruct.AlignOf.Should().Be(4);
+        _ = anonymousStruct.IsAnonymous.Should().BeTrue();
+        _ = anonymousStruct.Fields.Length.Should().Be(2);
 
         var anonymousField1 = anonymousStruct.Fields[0];
-        anonymousField1.Name.Should().BeEmpty();
-        anonymousField1.OffsetOf.Should().Be(0);
-        anonymousField1.Type.Name.Should().Be(anonymousStruct.Name + "_ANONYMOUS_0");
-        anonymousField1.Type.SizeOf.Should().Be(8);
-        anonymousField1.Type.AlignOf.Should().Be(4);
-        anonymousField1.Type.IsAnonymous.Should().BeTrue();
-        anonymousField1.Type.InnerType.Should().BeNull();
+        _ = anonymousField1.Name.Should().BeEmpty();
+        _ = anonymousField1.OffsetOf.Should().Be(0);
+        _ = anonymousField1.Type.Name.Should().Be(anonymousStruct.Name + "_ANONYMOUS_0");
+        _ = anonymousField1.Type.SizeOf.Should().Be(8);
+        _ = anonymousField1.Type.AlignOf.Should().Be(4);
+        _ = anonymousField1.Type.IsAnonymous.Should().BeTrue();
+        _ = anonymousField1.Type.InnerType.Should().BeNull();
 
         var anonymousField2 = anonymousStruct.Fields[1];
-        anonymousField2.Name.Should().BeEmpty();
-        anonymousField2.OffsetOf.Should().Be(8);
-        anonymousField2.Type.Name.Should().Be(anonymousStruct.Name + "_ANONYMOUS_1");
-        anonymousField2.Type.SizeOf.Should().Be(8);
-        anonymousField2.Type.AlignOf.Should().Be(4);
-        anonymousField2.Type.IsAnonymous.Should().BeTrue();
-        anonymousField2.Type.InnerType.Should().BeNull();
+        _ = anonymousField2.Name.Should().BeEmpty();
+        _ = anonymousField2.OffsetOf.Should().Be(8);
+        _ = anonymousField2.Type.Name.Should().Be(anonymousStruct.Name + "_ANONYMOUS_1");
+        _ = anonymousField2.Type.SizeOf.Should().Be(8);
+        _ = anonymousField2.Type.AlignOf.Should().Be(4);
+        _ = anonymousField2.Type.IsAnonymous.Should().BeTrue();
+        _ = anonymousField2.Type.InnerType.Should().BeNull();
 
         var nestedAnonymousStruct1 = ffi.GetRecord(anonymousField1.Type.Name);
-        nestedAnonymousStruct1.IsStruct.Should().BeTrue();
-        nestedAnonymousStruct1.IsUnion.Should().BeFalse();
-        nestedAnonymousStruct1.SizeOf.Should().Be(8);
-        nestedAnonymousStruct1.AlignOf.Should().Be(4);
-        nestedAnonymousStruct1.IsAnonymous.Should().BeTrue();
-        nestedAnonymousStruct1.Fields.Length.Should().Be(2);
+        _ = nestedAnonymousStruct1.IsStruct.Should().BeTrue();
+        _ = nestedAnonymousStruct1.IsUnion.Should().BeFalse();
+        _ = nestedAnonymousStruct1.SizeOf.Should().Be(8);
+        _ = nestedAnonymousStruct1.AlignOf.Should().Be(4);
+        _ = nestedAnonymousStruct1.IsAnonymous.Should().BeTrue();
+        _ = nestedAnonymousStruct1.Fields.Length.Should().Be(2);
 
         var nestedAnonymousStruct1Field1 = nestedAnonymousStruct1.Fields[0];
-        nestedAnonymousStruct1Field1.Name.Should().Be("a");
+        _ = nestedAnonymousStruct1Field1.Name.Should().Be("a");
         nestedAnonymousStruct1Field1.Type.Should().BeChar();
-        nestedAnonymousStruct1Field1.OffsetOf.Should().Be(0);
+        _ = nestedAnonymousStruct1Field1.OffsetOf.Should().Be(0);
 
         var nestedAnonymousStruct1Field2 = nestedAnonymousStruct1.Fields[1];
-        nestedAnonymousStruct1Field2.Name.Should().Be("b");
+        _ = nestedAnonymousStruct1Field2.Name.Should().Be("b");
         nestedAnonymousStruct1Field2.Type.Should().BeInt();
-        nestedAnonymousStruct1Field2.OffsetOf.Should().Be(4);
+        _ = nestedAnonymousStruct1Field2.OffsetOf.Should().Be(4);
 
         var nestedAnonymousStruct2 = ffi.GetRecord(anonymousField2.Type.Name);
-        nestedAnonymousStruct2.IsStruct.Should().BeTrue();
-        nestedAnonymousStruct2.IsUnion.Should().BeFalse();
-        nestedAnonymousStruct2.SizeOf.Should().Be(8);
-        nestedAnonymousStruct2.AlignOf.Should().Be(4);
-        nestedAnonymousStruct2.IsAnonymous.Should().BeTrue();
-        nestedAnonymousStruct2.Fields.Length.Should().Be(2);
+        _ = nestedAnonymousStruct2.IsStruct.Should().BeTrue();
+        _ = nestedAnonymousStruct2.IsUnion.Should().BeFalse();
+        _ = nestedAnonymousStruct2.SizeOf.Should().Be(8);
+        _ = nestedAnonymousStruct2.AlignOf.Should().Be(4);
+        _ = nestedAnonymousStruct2.IsAnonymous.Should().BeTrue();
+        _ = nestedAnonymousStruct2.Fields.Length.Should().Be(2);
 
         var nestedAnonymousStruct2Field1 = nestedAnonymousStruct2.Fields[0];
-        nestedAnonymousStruct2Field1.Name.Should().Be("c");
+        _ = nestedAnonymousStruct2Field1.Name.Should().Be("c");
         nestedAnonymousStruct2Field1.Type.Should().BeChar();
-        nestedAnonymousStruct2Field1.OffsetOf.Should().Be(0);
+        _ = nestedAnonymousStruct2Field1.OffsetOf.Should().Be(0);
 
         var nestedAnonymousStruct2Field2 = nestedAnonymousStruct2.Fields[1];
-        nestedAnonymousStruct2Field2.Name.Should().Be("d");
+        _ = nestedAnonymousStruct2Field2.Name.Should().Be("d");
         nestedAnonymousStruct2Field2.Type.Should().BeInt();
-        nestedAnonymousStruct2Field2.OffsetOf.Should().Be(4);
+        _ = nestedAnonymousStruct2Field2.OffsetOf.Should().Be(4);
     }
 }
