@@ -1,10 +1,6 @@
 // Copyright (c) Bottlenose Labs Inc. (https://github.com/bottlenoselabs). All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the Git repository root directory for full license information.
 
-using c2ffi.Tests.Library.Models;
-using FluentAssertions;
-using Xunit;
-
 #pragma warning disable CA1707
 
 namespace c2ffi.Tests.EndToEnd.Merge.Functions.function_int_params_int;
@@ -25,18 +21,18 @@ public class Test : MergeFfisTest
     private void FfiFunctionExists(CTestFfiCrossPlatform ffi)
     {
         var function = ffi.GetFunction(FunctionName);
-        function.CallingConvention.Should().Be("cdecl");
+        _ = function.CallingConvention.Should().Be("cdecl");
 
         var returnType = function.ReturnType;
-        returnType.Name.Should().Be("int");
-        returnType.NodeKind.Should().Be("primitive");
-        returnType.SizeOf.Should().Be(4);
-        returnType.AlignOf.Should().Be(4);
-        returnType.InnerType.Should().BeNull();
+        _ = returnType.Name.Should().Be("int");
+        _ = returnType.NodeKind.Should().Be("primitive");
+        _ = returnType.SizeOf.Should().Be(4);
+        _ = returnType.AlignOf.Should().Be(4);
+        _ = returnType.InnerType.Should().BeNull();
 
-        function.Parameters.Length.Should().Be(1);
+        _ = function.Parameters.Length.Should().Be(1);
         var parameter = function.Parameters[0];
-        parameter.Name.Should().Be("a");
-        parameter.Type.Name.Should().Be("int");
+        _ = parameter.Name.Should().Be("a");
+        _ = parameter.Type.Name.Should().Be("int");
     }
 }
