@@ -1,10 +1,6 @@
 // Copyright (c) Bottlenose Labs Inc. (https://github.com/bottlenoselabs). All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the Git repository root directory for full license information.
 
-using c2ffi.Tests.Library.Models;
-using FluentAssertions;
-using Xunit;
-
 #pragma warning disable CA1707
 
 namespace c2ffi.Tests.EndToEnd.Merge.Functions.function_uint64_params_uint8_uint16_uint32;
@@ -25,27 +21,27 @@ public class Test : MergeFfisTest
     private static void FfiFunctionExists(CTestFfiCrossPlatform ffi)
     {
         var function = ffi.GetFunction(FunctionName);
-        function.CallingConvention.Should().Be("cdecl");
+        _ = function.CallingConvention.Should().Be("cdecl");
 
         var returnType = function.ReturnType;
-        returnType.Name.Should().Be("uint64_t");
-        returnType.NodeKind.Should().Be("typealias");
-        returnType.SizeOf.Should().Be(8);
-        returnType.AlignOf.Should().Be(8);
-        returnType.InnerType.Should().NotBeNull();
+        _ = returnType.Name.Should().Be("uint64_t");
+        _ = returnType.NodeKind.Should().Be("typealias");
+        _ = returnType.SizeOf.Should().Be(8);
+        _ = returnType.AlignOf.Should().Be(8);
+        _ = returnType.InnerType.Should().NotBeNull();
 
-        function.Parameters.Length.Should().Be(3);
+        _ = function.Parameters.Length.Should().Be(3);
 
         var parameter1 = function.Parameters[0];
-        parameter1.Name.Should().Be("a");
-        parameter1.Type.Name.Should().Be("uint8_t");
+        _ = parameter1.Name.Should().Be("a");
+        _ = parameter1.Type.Name.Should().Be("uint8_t");
 
         var parameter2 = function.Parameters[1];
-        parameter2.Name.Should().Be("b");
-        parameter2.Type.Name.Should().Be("uint16_t");
+        _ = parameter2.Name.Should().Be("b");
+        _ = parameter2.Type.Name.Should().Be("uint16_t");
 
         var parameter3 = function.Parameters[2];
-        parameter3.Name.Should().Be("c");
-        parameter3.Type.Name.Should().Be("uint32_t");
+        _ = parameter3.Name.Should().Be("c");
+        _ = parameter3.Type.Name.Should().Be("uint32_t");
     }
 }
