@@ -1,10 +1,6 @@
 // Copyright (c) Bottlenose Labs Inc. (https://github.com/bottlenoselabs). All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the Git repository root directory for full license information.
 
-using c2ffi.Tests.Library.Models;
-using FluentAssertions;
-using Xunit;
-
 #pragma warning disable CA1707
 
 namespace c2ffi.Tests.EndToEnd.Merge.Functions.function_void;
@@ -25,15 +21,15 @@ public class Test : MergeFfisTest
     private void FfiFunctionExists(CTestFfiCrossPlatform ffi)
     {
         var function = ffi.GetFunction(FunctionName);
-        function.CallingConvention.Should().Be("cdecl");
+        _ = function.CallingConvention.Should().Be("cdecl");
 
         var returnType = function.ReturnType;
-        returnType.Name.Should().Be("void");
-        returnType.NodeKind.Should().Be("primitive");
-        returnType.SizeOf.Should().BeNull();
-        returnType.AlignOf.Should().BeNull();
-        returnType.InnerType.Should().BeNull();
+        _ = returnType.Name.Should().Be("void");
+        _ = returnType.NodeKind.Should().Be("primitive");
+        _ = returnType.SizeOf.Should().BeNull();
+        _ = returnType.AlignOf.Should().BeNull();
+        _ = returnType.InnerType.Should().BeNull();
 
-        function.Parameters.Should().BeEmpty();
+        _ = function.Parameters.Should().BeEmpty();
     }
 }
