@@ -1,10 +1,6 @@
 // Copyright (c) Bottlenose Labs Inc. (https://github.com/bottlenoselabs). All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the Git repository root directory for full license information.
 
-using c2ffi.Tests.Library.Models;
-using FluentAssertions;
-using Xunit;
-
 #pragma warning disable CA1308
 #pragma warning disable CA1707
 
@@ -25,9 +21,9 @@ public class Test : MergeFfisTest
     private void FfiMacroObjectExists(CTestFfiCrossPlatform ffi)
     {
         var macroObject = ffi.GetMacroObject(MacroObjectName);
-        macroObject.Name.Should().Be(MacroObjectName);
-        macroObject.Value.Should().Be("42");
-        macroObject.Type.Name.Should().Be("int");
-        macroObject.Type.InnerType.Should().BeNull();
+        _ = macroObject.Name.Should().Be(MacroObjectName);
+        _ = macroObject.Value.Should().Be("42");
+        _ = macroObject.Type.Name.Should().Be("int");
+        _ = macroObject.Type.InnerType.Should().BeNull();
     }
 }
