@@ -1,10 +1,6 @@
 // Copyright (c) Bottlenose Labs Inc. (https://github.com/bottlenoselabs). All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the Git repository root directory for full license information.
 
-using c2ffi.Tests.Library.Models;
-using FluentAssertions;
-using Xunit;
-
 #pragma warning disable CA1707
 
 namespace c2ffi.Tests.EndToEnd.Merge.Structs.struct_int;
@@ -26,15 +22,15 @@ public class Test : MergeFfisTest
     {
         const string name = $"struct {StructName}";
         var record = ffi.GetRecord(name);
-        record.Name.Should().Be(name);
-        record.IsStruct.Should().Be(true);
-        record.IsUnion.Should().Be(false);
+        _ = record.Name.Should().Be(name);
+        _ = record.IsStruct.Should().Be(true);
+        _ = record.IsUnion.Should().Be(false);
 
-        record.Fields.Length.Should().Be(1);
+        _ = record.Fields.Length.Should().Be(1);
 
         var field = record.Fields[0];
-        field.Name.Should().Be("a");
-        field.OffsetOf.Should().Be(0);
+        _ = field.Name.Should().Be("a");
+        _ = field.OffsetOf.Should().Be(0);
         field.Type.Should().BeInt();
     }
 }
