@@ -1,10 +1,6 @@
 // Copyright (c) Bottlenose Labs Inc. (https://github.com/bottlenoselabs). All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the Git repository root directory for full license information.
 
-using c2ffi.Tests.Library.Models;
-using FluentAssertions;
-using Xunit;
-
 #pragma warning disable CA1707
 
 namespace c2ffi.Tests.EndToEnd.Merge.Variables.variable_ignored;
@@ -37,7 +33,7 @@ public class Test : MergeFfisTest
         foreach (var name in names)
         {
             var variable = ffi.TryGetVariable(name);
-            variable.Should().NotBeNull();
+            _ = variable.Should().NotBeNull();
         }
     }
 
@@ -46,7 +42,7 @@ public class Test : MergeFfisTest
         foreach (var name in names)
         {
             var variable = ffi.TryGetVariable(name);
-            variable.Should().BeNull();
+            _ = variable.Should().BeNull();
         }
     }
 }
