@@ -1,10 +1,6 @@
 // Copyright (c) Bottlenose Labs Inc. (https://github.com/bottlenoselabs). All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the Git repository root directory for full license information.
 
-using c2ffi.Tests.Library.Models;
-using FluentAssertions;
-using Xunit;
-
 #pragma warning disable CA1707
 
 namespace c2ffi.Tests.EndToEnd.Merge.Variables.variable_int;
@@ -24,7 +20,7 @@ public class Test : MergeFfisTest
     private void FfiVariableExists(CTestFfiCrossPlatform ffi)
     {
         var variable = ffi.GetVariable(VariableName);
-        variable.Name.Should().Be(VariableName);
-        variable.TypeName.Should().Be("int");
+        _ = variable.Name.Should().Be(VariableName);
+        _ = variable.TypeName.Should().Be("int");
     }
 }
